Fix a bug in the last commit.  4.x improperly tries to add the children of a PCI bus during the probe phase.  This doesn't mix well with 5.x's bus code which adds the children in the attach phase.  Clean up pci_probe().  The code, unfortunately, is still quite fragile.
Cleanup pci_read_device().

Bring in pcib_*() routines from 5.x into pcisupport.c.  Add support for
bridges which do not support particular PCI mapping types (from FreeBSD 5.x).

Fix a bug in the pcib_route_interrupt() method.

+++ sys/bus/pci/pci_private.h
@@ -23,13 +23,13 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF
  * THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  *
- * $DragonFly: src/sys/bus/pci/pci_private.h,v 1.2 2004/02/21 06:37:05 dillon Exp $
+ * $DragonFly: src/sys/bus/pci/pci_private.h,v 1.3 2004/02/21 09:16:27 dillon Exp $
  *
  */
 
 extern devclass_t	pci_devclass;
 
-struct pci_devinfo 	*pci_read_device(device_t, int, int, int, int);
+struct pci_devinfo 	*pci_read_device(device_t, int, int, int, size_t);
 struct resource_list	*pci_get_resource_list (device_t dev, device_t child);
 static struct resource *pci_alloc_resource(device_t dev, device_t child,
 				int type, int *rid, u_long start, u_long end,
